requirements.txt
flask==3.1.0
uvicorn==0.34.0
requests==2.32.3
pyyaml==6.0.2